[TASK] Move shared functionality of textroles to common abstract class
